requirements.txt
Flask==3.0.3
requests==2.31.0
matplotlib==3.8.4
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
